// web/src/api/generated/model/depreciationMethod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type DepreciationMethod = typeof DepreciationMethod[keyof typeof DepreciationMethod];


export const DepreciationMethod = {
  StraightLine: 'StraightLine',
  DecliningBalance: 'DecliningBalance',
} as const;